+++ src/main/java/GDG4th/personaChat/auth/presentation/dto/LoginRequest.java
@@ -1,0 +1,9 @@
+package GDG4th.personaChat.auth.presentation.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginRequest(
+        @NotBlank String email,
+        @NotBlank String password
+) {
+}
